requirements.txt
fastapi
uvicorn
pandas
numpy
pvlib
scikit-learn==1.6.1
joblib
requests
python-dotenv
